refactor: RequestHandler 클래스 분리
